Switch to using Cobra CLI
